+++ src/imp/auth_imp.js
@@ -1,4 +1,6 @@
-import { crouton_thrift } from '../platform_abstraction_layer'; // eslint-disable-line camelcase
+// eslint-disable-next-line camelcase
+import { crouton_thrift } from '../platform_abstraction_layer';
+
 let proto = require('./generated_proto/collector_pb.js');
 
 export default class AuthImp {