+++ src/imp/platform/browser/platform_browser.js
@@ -7,10 +7,10 @@ const kCookieTimeToLiveSeconds = 7 * 24 * 60 * 60;
 
 let nowMicrosImp = (function () {
     // Is a hi-res timer available?
-    if (window.performance &&
-        window.performance.now &&
-        window.performance.timing &&
-        window.performance.timing.navigationStart) {
+    if (window.performance
+        && window.performance.now
+        && window.performance.timing
+        && window.performance.timing.navigationStart) {
         let start = performance.timing.navigationStart;
         return function () {
             return Math.floor((start + performance.now()) * 1000.0);
@@ -23,7 +23,6 @@ let nowMicrosImp = (function () {
 }());
 
 class PlatformBrowser {
-
     name() {
         return 'browser';
     }
@@ -56,13 +55,18 @@ class PlatformBrowser {
     }
 
     _generateLongUUID() {
+        /* eslint-disable no-bitwise */
         let p0 = `00000000${Math.abs((Math.random() * 0xFFFFFFFF) | 0).toString(16)}`.substr(-8);
         let p1 = `00000000${Math.abs((Math.random() * 0xFFFFFFFF) | 0).toString(16)}`.substr(-8);
         return `${p0}${p1}`;
+        /* eslint-enable no-bitwise */
     }
 
     onBeforeExit(...args) {
-        if (window) {
+        // This will result in the final report not being made in non-browser
+        // environments like React Native. Flush should be called explicitly in
+        // those environments
+        if (util.isBrowser()) {
             window.addEventListener('beforeunload', ...args);
         }
     }
@@ -92,6 +96,7 @@ class PlatformBrowser {
             PlatformBrowser.initGlobalTracer(lib, tracerOpts);
         }
     }
+
     static initGlobalTracer(lib, opts) {
         if (typeof window !== 'object') {
             return;
@@ -99,7 +104,7 @@ class PlatformBrowser {
         if (typeof window.opentracing !== 'object') {
             return;
         }
-        opentracing.initGlobalTracer(new lib.Tracer(opts));  // eslint-disable-line no-undef
+        opentracing.initGlobalTracer(new lib.Tracer(opts)); // eslint-disable-line no-undef
     }
 
     tracerTags() {