+++ src/imp/platform/node/transport_httpproto.js
@@ -1,5 +1,6 @@
 import * as https from 'https';
 import * as http from 'http';
+
 let proto = require('../../generated_proto/collector_pb.js');
 
 const kMaxDetailedErrorFrequencyMs = 30000;