+++ src/imp/span_imp.js
@@ -1,16 +1,16 @@
-import * as coerce from './coerce.js';
+import * as opentracing from 'opentracing';
+import * as coerce from './coerce';
 import * as constants from '../constants';
 import _each from '../_each';
-import * as opentracing from 'opentracing';
 import { crouton_thrift } from '../platform_abstraction_layer'; // eslint-disable-line camelcase
 import LogRecordImp from './log_record_imp'; // eslint-disable-line camelcase
-import util from './util/util.js';
+import util from './util/util';
+
 let converter = require('hex2dec');
-let proto = require('./generated_proto/collector_pb.js');
-let googleProtobufTimestampPB = require('google-protobuf/google/protobuf/timestamp_pb.js');
+let googleProtobufTimestampPB = require('google-protobuf/google/protobuf/timestamp_pb');
+let proto = require('./generated_proto/collector_pb');
 
 export default class SpanImp extends opentracing.Span {
-
     // ---------------------------------------------------------------------- //
     // opentracing.Span SPI
     // ---------------------------------------------------------------------- //
@@ -25,6 +25,14 @@ export default class SpanImp extends opentracing.Span {
 
     _setOperationName(name) {
         this._operationName = `${name}`;
+    }
+
+    _setBaggageItem(key, value) {
+        this._ctx.setBaggageItem(key, value);
+    }
+
+    _getBaggageItem(key) {
+        return this._ctx.getBaggageItem(key);
     }
 
     _addTags(keyValuePairs) {
@@ -42,15 +50,16 @@ export default class SpanImp extends opentracing.Span {
             return;
         }
 
-        let tsMicros = timestamp ?
-            (timestamp * 1000) :
-            self._tracerImp._platform.nowMicros();
+        let tsMicros = timestamp
+            ? (timestamp * 1000)
+            : self._tracerImp._platform.nowMicros();
 
         let record = new LogRecordImp(
             self._tracerImp.getLogFieldKeyHardLimit(),
             self._tracerImp.getLogFieldValueHardLimit(),
             tsMicros,
-            keyValuePairs);
+            keyValuePairs,
+        );
         self._log_records = self._log_records || [];
         self._log_records.push(record);
         self._tracerImp.emit('log_added', record);
@@ -67,7 +76,7 @@ export default class SpanImp extends opentracing.Span {
     constructor(tracer, name, spanContext) {
         super();
 
-        console.assert(typeof tracer === 'object', 'Invalid runtime');  // eslint-disable-line no-console
+        console.assert(typeof tracer === 'object', 'Invalid runtime'); // eslint-disable-line no-console
 
         this._tracerImp = tracer;
         this._ctx = spanContext;
@@ -154,8 +163,8 @@ export default class SpanImp extends opentracing.Span {
      * Finishes the span.
      *
      * @param  {Number} finishTime
-     *         	Optional Unix timestamp in milliseconds setting an explicit
-     *         	finish time for the span.
+     *         Optional Unix timestamp in milliseconds setting an explicit
+     *         finish time for the span.
      */
     end(finishTime) {
         // Ensure a single span is not recorded multiple times
@@ -230,9 +239,8 @@ export default class SpanImp extends opentracing.Span {
         spanProto.setOperationName(this._operationName);
 
         let startTimestamp = new googleProtobufTimestampPB.Timestamp();
-        let startMillis = Math.floor(this._beginMicros / 1000);
-        let startSeconds = Math.floor(startMillis / 1000);
-        let startNanos = (startMillis % 1000) * 1000000;
+        let startSeconds = Math.floor(this._beginMicros / 1000000);
+        let startNanos = (this._beginMicros % 1000000) * 1000;
         startTimestamp.setSeconds(startSeconds);
         startTimestamp.setNanos(startNanos);
         spanProto.setStartTimestamp(startTimestamp);
@@ -247,7 +255,7 @@ export default class SpanImp extends opentracing.Span {
         });
         spanProto.setLogsList(logs);
 
-        let parentSpanGUID = undefined;
+        let parentSpanGUID;
         let tags = [];
         _each(this._tags, (value, key) => {
             let strValue = coerce.toString(value);